+++ include/monolish_vml_float.hpp
@@ -265,7 +265,8 @@ void sign(const vector<float> &a, vector<float> &y);
 void reciprocal(const vector<float> &a, vector<float> &y);
 
 /**
- * @brief Create a new vector with greatest elements of two vectors (y[0:N] = max(a[0:N], b[0]:N]))
+ * @brief Create a new vector with greatest elements of two vectors (y[0:N] =
+ *max(a[0:N], b[0]:N]))
  * @param a single precision monolish vector (size N)
  * @param b single precision monolish vector (size N)
  * @param y single precision monolish vector (size N)
@@ -277,7 +278,8 @@ void reciprocal(const vector<float> &a, vector<float> &y);
 void max(const vector<float> &a, const vector<float> &b, vector<float> &y);
 
 /**
- * @brief Create a new vector with smallest elements of two vectors (y[0:N] = max(a[0:N], b[0]:N]))
+ * @brief Create a new vector with smallest elements of two vectors (y[0:N] =
+ *max(a[0:N], b[0]:N]))
  * @param a single precision monolish vector (size N)
  * @param b single precision monolish vector (size N)
  * @param y single precision monolish vector (size N)
@@ -289,7 +291,8 @@ void max(const vector<float> &a, const vector<float> &b, vector<float> &y);
 void min(const vector<float> &a, const vector<float> &b, vector<float> &y);
 
 /**
- * @brief Finds the greatest element in single precision vector (max_element(y[0:N]))
+ * @brief Finds the greatest element in single precision vector
+ *(max_element(y[0:N]))
  * @param y single precision monolish vector (size N)
  * @return greatest value
  * @note
@@ -300,7 +303,8 @@ void min(const vector<float> &a, const vector<float> &b, vector<float> &y);
 float max(const vector<float> &y);
 
 /**
- * @brief Finds the smallest element in single precision vector (max_element(y[0:N]))
+ * @brief Finds the smallest element in single precision vector
+ *(max_element(y[0:N]))
  * @param y single precision monolish vector (size N)
  * @return smallest value
  * @note
@@ -610,7 +614,8 @@ void sign(const matrix::Dense<float> &A, matrix::Dense<float> &C);
 void reciprocal(const matrix::Dense<float> &A, matrix::Dense<float> &C);
 
 /**
- * @brief Create a new matrix with greatest elements of two matrices (C[0:nnz] = A(a[0:nnz], B[0]:nnz]))
+ * @brief Create a new matrix with greatest elements of two matrices (C[0:nnz] =
+ *A(a[0:nnz], B[0]:nnz]))
  * @param A single precision monolish Dense matrix (size M x N)
  * @param B single precision monolish Dense matrix (size M x N)
  * @param C single precision monolish Dense matrix (size M x N)
@@ -619,10 +624,12 @@ void reciprocal(const matrix::Dense<float> &A, matrix::Dense<float> &C);
  * - Multi-threading: true
  * - GPU acceleration: true
  **/
-void max(const matrix::Dense<float> &A, const matrix::Dense<float> &B, matrix::Dense<float> &C);
+void max(const matrix::Dense<float> &A, const matrix::Dense<float> &B,
+         matrix::Dense<float> &C);
 
 /**
- * @brief Create a new matrix with smallest elements of two matrices (C[0:nnz] = A(a[0:nnz], B[0]:nnz]))
+ * @brief Create a new matrix with smallest elements of two matrices (C[0:nnz] =
+ *A(a[0:nnz], B[0]:nnz]))
  * @param A single precision monolish Dense matrix (size M x N)
  * @param B single precision monolish Dense matrix (size M x N)
  * @param C single precision monolish Dense matrix (size M x N)
@@ -631,10 +638,12 @@ void max(const matrix::Dense<float> &A, const matrix::Dense<float> &B, matrix::D
  * - Multi-threading: true
  * - GPU acceleration: true
  **/
-void min(const matrix::Dense<float> &A, const matrix::Dense<float> &B, matrix::Dense<float> &C);
+void min(const matrix::Dense<float> &A, const matrix::Dense<float> &B,
+         matrix::Dense<float> &C);
 
 /**
- * @brief Finds the greatest element in single precision Dense matrices (max_element(C[0:nnz]))
+ * @brief Finds the greatest element in single precision Dense matrices
+ *(max_element(C[0:nnz]))
  * @param C single precision monolish Dense matrix (size M x N)
  * @return greatest value
  * @note
@@ -644,7 +653,8 @@ void min(const matrix::Dense<float> &A, const matrix::Dense<float> &B, matrix::D
  **/
 float max(const matrix::Dense<float> &C);
 /**
- * @brief Finds the greatest element in single precision Dense matrices (max_element(C[0:nnz]))
+ * @brief Finds the greatest element in single precision Dense matrices
+ *(max_element(C[0:nnz]))
  * @param C single precision monolish Dense matrix (size M x N)
  * @return greatest value
  * @note
@@ -739,8 +749,7 @@ void div(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
  * @warning
  * A and C must be same non-zero structure (dont check in this function)
  */
-void add(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void add(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief single precision scalar and Dence matrix subtraction (C[i][j] =
@@ -756,8 +765,7 @@ void add(const matrix::CRS<float> &A, const float alpha,
  * @warning
  * A and C must be same non-zero structure (dont check in this function)
  */
-void sub(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void sub(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief single precision scalar and Dence matrix multiplication (C[i][j] =
@@ -773,8 +781,7 @@ void sub(const matrix::CRS<float> &A, const float alpha,
  * @warning
  * A and C must be same non-zero structure (dont check in this function)
  */
-void mul(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void mul(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief single precision scalar and Dence matrix division (C[i][j] = A[i][j] +
@@ -790,8 +797,7 @@ void mul(const matrix::CRS<float> &A, const float alpha,
  * @warning
  * A and C must be same non-zero structure (dont check in this function)
  */
-void div(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void div(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief power to single precision CRS matrix elements by single precision CRS
@@ -822,8 +828,7 @@ void pow(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
  * @warning
  * A and C must be same non-zero structure (dont check in this function)
  **/
-void pow(const matrix::CRS<float> &A, const float alpha,
-         matrix::CRS<float> &C);
+void pow(const matrix::CRS<float> &A, const float alpha, matrix::CRS<float> &C);
 
 /**
  * @brief sqrt to single precision CRS matrix elements (C[0:N] = sqrt(A[0:N]))
@@ -1009,7 +1014,8 @@ void sign(const matrix::CRS<float> &A, matrix::CRS<float> &C);
 void reciprocal(const matrix::CRS<float> &A, matrix::CRS<float> &C);
 
 /**
- * @brief Create a new matrix with greatest elements of two matrices (C[0:nnz] = A(a[0:nnz], B[0]:nnz]))
+ * @brief Create a new matrix with greatest elements of two matrices (C[0:nnz] =
+ *A(a[0:nnz], B[0]:nnz]))
  * @param A single precision monolish CRS matrix (size M x N)
  * @param B single precision monolish CRS matrix (size M x N)
  * @param C single precision monolish CRS matrix (size M x N)
@@ -1020,10 +1026,12 @@ void reciprocal(const matrix::CRS<float> &A, matrix::CRS<float> &C);
  * @warning
  * A, B and C must be same non-zero structure (dont check in this function)
  **/
-void max(const matrix::CRS<float> &A, const matrix::CRS<float> &B, matrix::CRS<float> &C);
+void max(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
+         matrix::CRS<float> &C);
 
 /**
- * @brief Create a new matrix with smallest elements of two matrices (C[0:nnz] = A(a[0:nnz], B[0]:nnz]))
+ * @brief Create a new matrix with smallest elements of two matrices (C[0:nnz] =
+ *A(a[0:nnz], B[0]:nnz]))
  * @param A single precision monolish CRS matrix (size M x N)
  * @param B single precision monolish CRS matrix (size M x N)
  * @param C single precision monolish CRS matrix (size M x N)
@@ -1034,10 +1042,12 @@ void max(const matrix::CRS<float> &A, const matrix::CRS<float> &B, matrix::CRS<f
  * @warning
  * A, B and C must be same non-zero structure (dont check in this function)
  **/
-void min(const matrix::CRS<float> &A, const matrix::CRS<float> &B, matrix::CRS<float> &C);
+void min(const matrix::CRS<float> &A, const matrix::CRS<float> &B,
+         matrix::CRS<float> &C);
 
 /**
- * @brief Finds the greatest element in single precision CRS matrices (max_element(C[0:nnz]))
+ * @brief Finds the greatest element in single precision CRS matrices
+ *(max_element(C[0:nnz]))
  * @param C single precision monolish CRS matrix (size M x N)
  * @return greatest value
  * @note
@@ -1047,7 +1057,8 @@ void min(const matrix::CRS<float> &A, const matrix::CRS<float> &B, matrix::CRS<f
  **/
 float max(const matrix::CRS<float> &C);
 /**
- * @brief Finds the greatest element in single precision CRS matrices (max_element(C[0:nnz]))
+ * @brief Finds the greatest element in single precision CRS matrices
+ *(max_element(C[0:nnz]))
  * @param C single precision monolish CRS matrix (size M x N)
  * @return greatest value
  * @note