+++ src/internal/math/double_math.cpp
@@ -407,12 +407,10 @@ double vmax(const size_t N, const double *y, bool gpu_status) {
   if (gpu_status == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for reduction(max : max)
-    for( size_t i=0;i<N; i++)
-    {
-        if(y[i] > max)
-        {
-            max = y[i];  
-        }
+    for (size_t i = 0; i < N; i++) {
+      if (y[i] > max) {
+        max = y[i];
+      }
     }
 #else
     throw std::runtime_error(
@@ -420,11 +418,9 @@ double vmax(const size_t N, const double *y, bool gpu_status) {
 #endif
   } else {
 #pragma omp parallel for reduction(max : max)
-    for( size_t i=0;i<N; i++)
-    {
-      if(y[i] > max)
-      {
-        max = y[i];  
+    for (size_t i = 0; i < N; i++) {
+      if (y[i] > max) {
+        max = y[i];
       }
     }
   }
@@ -432,26 +428,25 @@ double vmax(const size_t N, const double *y, bool gpu_status) {
   return max;
 }
 
-void vmax(const size_t N, const double *a, const double *b, double *y, bool gpu_status) {
+void vmax(const size_t N, const double *a, const double *b, double *y,
+          bool gpu_status) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
   if (gpu_status == true) {
 #if MONOLISH_USE_GPU
-#pragma omp target teams distribute parallel for 
-    for( size_t i=0;i<N; i++)
-    {
-      y[i] = std::max(a[i], b[i]);  
+#pragma omp target teams distribute parallel for
+    for (size_t i = 0; i < N; i++) {
+      y[i] = std::max(a[i], b[i]);
     }
 #else
     throw std::runtime_error(
         "error USE_GPU is false, but get_device_mem_stat() == true");
 #endif
   } else {
-#pragma omp parallel for 
-    for( size_t i=0;i<N; i++)
-    {
-      y[i] = std::max(a[i], b[i]);  
+#pragma omp parallel for
+    for (size_t i = 0; i < N; i++) {
+      y[i] = std::max(a[i], b[i]);
     }
   }
   logger.func_out();
@@ -469,12 +464,10 @@ double vmin(const size_t N, const double *y, bool gpu_status) {
   if (gpu_status == true) {
 #if MONOLISH_USE_GPU
 #pragma omp target teams distribute parallel for reduction(min : min)
-    for( size_t i=0;i<N; i++)
-    {
-        if(y[i] < min)
-        {
-            min = y[i];  
-        }
+    for (size_t i = 0; i < N; i++) {
+      if (y[i] < min) {
+        min = y[i];
+      }
     }
 #else
     throw std::runtime_error(
@@ -482,37 +475,34 @@ double vmin(const size_t N, const double *y, bool gpu_status) {
 #endif
   } else {
 #pragma omp parallel for reduction(min : min)
-    for( size_t i=0;i<N; i++)
-    {
-      if(y[i] < min)
-      {
-        min = y[i];  
+    for (size_t i = 0; i < N; i++) {
+      if (y[i] < min) {
+        min = y[i];
       }
     }
   }
   logger.func_out();
   return min;
 }
-void vmin(const size_t N, const double *a, const double *b, double *y, bool gpu_status) {
+void vmin(const size_t N, const double *a, const double *b, double *y,
+          bool gpu_status) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
   if (gpu_status == true) {
 #if MONOLISH_USE_GPU
-#pragma omp target teams distribute parallel for 
-    for( size_t i=0;i<N; i++)
-    {
-      y[i] = std::min(a[i], b[i]);  
+#pragma omp target teams distribute parallel for
+    for (size_t i = 0; i < N; i++) {
+      y[i] = std::min(a[i], b[i]);
     }
 #else
     throw std::runtime_error(
         "error USE_GPU is false, but get_device_mem_stat() == true");
 #endif
   } else {
-#pragma omp parallel for 
-    for( size_t i=0;i<N; i++)
-    {
-      y[i] = std::min(a[i], b[i]);  
+#pragma omp parallel for
+    for (size_t i = 0; i < N; i++) {
+      y[i] = std::min(a[i], b[i]);
     }
   }
   logger.func_out();