+++ src/internal/monolish_internal.hpp
@@ -110,8 +110,10 @@ void vfloor(const size_t N, const double *a, double *y, bool gpu_status);
 void vsign(const size_t N, const double *a, double *y, bool gpu_status);
 double vmax(const size_t N, const double *y, bool gpu_status);
 double vmin(const size_t N, const double *y, bool gpu_status);
-void vmax(const size_t N, const double *a, const double *b, double *y, bool gpu_status);
-void vmin(const size_t N, const double *a, const double *b, double *y, bool gpu_status);
+void vmax(const size_t N, const double *a, const double *b, double *y,
+          bool gpu_status);
+void vmin(const size_t N, const double *a, const double *b, double *y,
+          bool gpu_status);
 
 void vsqrt(const size_t N, const float *a, float *y, bool gpu_status);
 void vpow(const size_t N, const float *a, const float *b, float *y,
@@ -131,8 +133,10 @@ void vfloor(const size_t N, const float *a, float *y, bool gpu_status);
 void vsign(const size_t N, const float *a, float *y, bool gpu_status);
 float vmax(const size_t N, const float *y, bool gpu_status);
 float vmin(const size_t N, const float *y, bool gpu_status);
-void vmax(const size_t N, const float *a, const float *b, float *y, bool gpu_status);
-void vmin(const size_t N, const float *a, const float *b, float *y, bool gpu_status);
+void vmax(const size_t N, const float *a, const float *b, float *y,
+          bool gpu_status);
+void vmin(const size_t N, const float *a, const float *b, float *y,
+          bool gpu_status);
 
 } // namespace internal
 } // namespace monolish