+++ src/vml/double_vector_math.cpp
@@ -274,7 +274,8 @@ void vml::reciprocal(const vector<double> &a, vector<double> &y) {
   logger.func_out();
 }
 
-void vml::max(const vector<double> &a, const vector<double> &b, vector<double> &y) {
+void vml::max(const vector<double> &a, const vector<double> &b,
+              vector<double> &y) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -282,11 +283,13 @@ void vml::max(const vector<double> &a, const vector<double> &b, vector<double> &
   if (a.size() != y.size() || a.size() != b.size()) {
     throw std::runtime_error("error vector size is not same");
   }
-  if (a.get_device_mem_stat() != y.get_device_mem_stat() || b.get_device_mem_stat() != y.get_device_mem_stat()) {
+  if (a.get_device_mem_stat() != y.get_device_mem_stat() ||
+      b.get_device_mem_stat() != y.get_device_mem_stat()) {
     throw std::runtime_error("error vector get_device_mem_stat() is not same");
   }
 
-  internal::vmax(y.size(), a.data(), b.data(), y.data(), y.get_device_mem_stat());
+  internal::vmax(y.size(), a.data(), b.data(), y.data(),
+                 y.get_device_mem_stat());
 
   logger.func_out();
 }
@@ -301,7 +304,8 @@ double vml::max(const vector<double> &y) {
   return max;
 }
 
-void vml::min(const vector<double> &a, const vector<double> &b, vector<double> &y) {
+void vml::min(const vector<double> &a, const vector<double> &b,
+              vector<double> &y) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
@@ -309,11 +313,13 @@ void vml::min(const vector<double> &a, const vector<double> &b, vector<double> &
   if (a.size() != y.size() || a.size() != b.size()) {
     throw std::runtime_error("error vector size is not same");
   }
-  if (a.get_device_mem_stat() != y.get_device_mem_stat() || b.get_device_mem_stat() != y.get_device_mem_stat()) {
+  if (a.get_device_mem_stat() != y.get_device_mem_stat() ||
+      b.get_device_mem_stat() != y.get_device_mem_stat()) {
     throw std::runtime_error("error vector get_device_mem_stat() is not same");
   }
 
-  internal::vmin(y.size(), a.data(), b.data(), y.data(), y.get_device_mem_stat());
+  internal::vmin(y.size(), a.data(), b.data(), y.data(),
+                 y.get_device_mem_stat());
 
   logger.func_out();
 }