+++ test/vector/vector_vml.cpp
@@ -17,6 +17,8 @@
 #include "vml/v_atanh.hpp"
 #include "vml/v_ceil.hpp"
 #include "vml/v_floor.hpp"
+#include "vml/v_max.hpp"
+#include "vml/v_min.hpp"
 #include "vml/v_reciprocal.hpp"
 #include "vml/v_sign.hpp"
 #include "vml/v_sin.hpp"
@@ -24,11 +26,9 @@
 #include "vml/v_sqrt.hpp"
 #include "vml/v_tan.hpp"
 #include "vml/v_tanh.hpp"
-#include "vml/vv_pow.hpp"
-#include "vml/v_max.hpp"
-#include "vml/v_min.hpp"
 #include "vml/vv_max.hpp"
 #include "vml/vv_min.hpp"
+#include "vml/vv_pow.hpp"
 
 int main(int argc, char **argv) {